+++ assets/js/theme/cart/CartUtils.js
@@ -79,6 +79,7 @@ export default class CartUtils {
 
           // update our cart model data for other apps and UI
           utils.api.cart.getCart({includeOptions: true}, (err, response) => {
+            console.log(response)
             window.TEAK.Utils.saveCartResponse(response);
           });
 
@@ -106,6 +107,7 @@ export default class CartUtils {
 
         // update our cart model data for other apps and UI
         utils.api.cart.getCart({includeOptions: true}, (err, response) => {
+          console.log(response)
           window.TEAK.Utils.saveCartResponse(response);
         });