+++ assets/js/theme/product/customizations/AddToCartModal.js
@@ -236,6 +236,7 @@ export default class AddToCartModal {
 
         // update our cart model data for other apps and UI
         utils.api.cart.getCart({includeOptions: true}, (err, response) => {
+          console.log(response)
           window.TEAK.Utils.saveCartResponse(response);
         });